requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.6.2
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.5.5
contourpy==1.1.0
cors==1.0.1
cycler==0.11.0
distlib==0.3.7
Django==4.1.13
django-3-jet==1.0.8
django-admin-material-dashboard==1.0.10
django-bootstrap4==24.3
django-cors-headers==4.3.1
django-crispy-forms==2.2
django-extensions==3.2.3
django-filter==23.5
django-jet-reboot==1.3.7
django-multiselectfield==0.1.12
django-python3-ldap==0.15.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.4.2
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.116.1
fastapi-cli==0.0.7
filelock==3.12.2
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
fonttools==4.40.0
future==0.18.3
gevent==23.7.0
geventhttpclient==2.0.9
graphviz==0.20.1
greenlet==2.0.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.6
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.4
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.10
ldap3==2.9.1
locust==2.15.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
msgpack==1.0.5
mysqlclient==2.2.0
numpy==1.25.0
openpyxl==3.1.2
orjson==3.10.18
outcome==1.3.0.post0
packaging==25.0
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.9.1
psutil==5.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==2.0.0
Pygments==2.19.1
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2023.3
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
roundrobin==0.0.4
scikit-learn==1.3.0
scipy==1.11.2
selenium==4.17.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.42
sqlparse==0.2.4
starlette==0.46.2
tenacity==9.1.2
threadpoolctl==3.2.0
tldextract==5.1.1
trio==0.24.0
trio-websocket==0.11.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.3
ujson==5.10.0
urllib3==2.0.4
virtualenv==20.24.1
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==2.3.6
wsproto==1.2.0
zope.event==5.0
zope.interface==6.0
zstandard==0.23.0
